+++ src/test/com/help/citrix/com/Test_JoinHelp_Page.java
@@ -7,7 +7,9 @@ import java.util.regex.Pattern;
 import org.openqa.selenium.WebDriver;
 //import org.openqa.selenium.WebElement;
 import org.testng.Assert;
+import org.testng.annotations.AfterTest;
 import org.testng.annotations.BeforeTest;
+import org.testng.annotations.Parameters;
 import org.testng.annotations.Test;
 //import com.help.citrix.Contact_Us_Page;
 //import com.help.citrix.GoToMeeting_Page;
@@ -19,20 +21,22 @@ import page.factory.helper.BrowserFactory;
 public class Test_JoinHelp_Page {
 	WebDriver driver;
 	JoinHelp_Page joinHelppg;
-	String browserName = "chrome";
-	String baseEnv = "ed1";
+	//String browserName = "firefox";
+	String baseEnv = "";
 	String baseUrl = "http://help" + baseEnv +".citrix.com";
 	String baseProduct ="/meeting/joinhelp";
 	String supportUrl = "/support";
 	
 	
 	@BeforeTest
-	public void setupJoinHelp_Page(){
-		System.out.println("BEFORE TEST METHOD");
-		driver = BrowserFactory.startBrowser(browserName, "about:blank");		
-		
-		driver.manage().deleteAllCookies();
-		driver.get(baseUrl+baseProduct);		
+	@Parameters("browser")
+	public void setupJoinHelp_Page(String browser){
+		System.out.println("@BeforeTest");
+		driver = BrowserFactory.startBrowser(browser, "about:blank");		
+		if(!browser.equalsIgnoreCase("Safari")){			
+			driver.manage().deleteAllCookies();
+		}
+		driver.get(baseUrl+baseProduct);	
 		
 	}
 	
@@ -40,8 +44,7 @@ public class Test_JoinHelp_Page {
 	public void verifyAttendeePage(){
 		
 		String article1Text = "I’m having trouble downloading the software.";
-		String article1Url = "http://support.citrixonline.com/Meeting/all_files/"
-				+ "G2M050002#Download?__";
+		String article1Url = "http://support.citrixonline.com/en_US/Meeting/all_files/G2M030004?__";
 		
 		String article2Text = "I’m having trouble connecting online using the Web App.";
 		String article2Url = "http://support.citrixonline.com/en_US/meeting/"
@@ -74,46 +77,65 @@ public class Test_JoinHelp_Page {
 		joinHelppg.imAttendee.click();
 	
 		//Verify the 1st Article
+		System.out.println("Verifying Attendee Article 1 Text:"+ joinHelppg.attendeeHelpLinks.get(0).getText());
 		Assert.assertEquals(joinHelppg.attendeeHelpLinks.get(0).getText(), article1Text);
+		System.out.println("Verifying Attendee Article 1 Url: " + joinHelppg.attendeeHelpLinks.get(0).getAttribute("href") );
 		Assert.assertTrue(joinHelppg.attendeeHelpLinks.get(0).getAttribute("href").
 					contains(article1Url));
 		joinHelppg.attendeeHelpLinks.get(0).click();
 		driver.navigate().back();
+		joinHelppg.imAttendee.click();
 		
 		//Verify the 2nd Article
+		System.out.println("Verifying Attendee Article 2 Text:"+ joinHelppg.attendeeHelpLinks.get(1).getText());
 		Assert.assertEquals(joinHelppg.attendeeHelpLinks.get(1).getText(), article2Text);
+		System.out.println("Verifying Attendee Article 2 Url: " + joinHelppg.attendeeHelpLinks.get(1).getAttribute("href") );
 		Assert.assertTrue(joinHelppg.attendeeHelpLinks.get(1).getAttribute("href").
 					contains(article2Url));
 		joinHelppg.attendeeHelpLinks.get(1).click();
 		driver.navigate().back();
+		joinHelppg.imAttendee.click();
 		
 		//Verify the 3rd Article
+		System.out.println("Verifying Attendee Article 3 Text:"+ joinHelppg.attendeeHelpLinks.get(2).getText());
 		Assert.assertEquals(joinHelppg.attendeeHelpLinks.get(2).getText(), article3Text);
+		System.out.println("Verifying Attendee Article 3 Url: " + joinHelppg.attendeeHelpLinks.get(2).getAttribute("href") );
 		Assert.assertTrue(joinHelppg.attendeeHelpLinks.get(2).getAttribute("href").
 					contains(article3Url));
 		joinHelppg.attendeeHelpLinks.get(2).click();
 		driver.navigate().back();
+		joinHelppg.imAttendee.click();
 		
 		//Verify the 4th Article
+		System.out.println("Verifying Attendee Article 4 Text:"+ joinHelppg.attendeeHelpLinks.get(3).getText());
 		Assert.assertEquals(joinHelppg.attendeeHelpLinks.get(3).getText(), article4Text);
+		System.out.println("Verifying Attendee Article 4 Url: " + joinHelppg.attendeeHelpLinks.get(3).getAttribute("href") );
 		Assert.assertTrue(joinHelppg.attendeeHelpLinks.get(3).getAttribute("href").
 					contains(article4Url));
 		joinHelppg.attendeeHelpLinks.get(3).click();
 		driver.navigate().back();
+		joinHelppg.imAttendee.click();
 				
 		//Verify the 5th Article
+		System.out.println("Verifying Attendee Article 5 Text:"+ joinHelppg.attendeeHelpLinks.get(4).getText());
 		Assert.assertEquals(joinHelppg.attendeeHelpLinks.get(4).getText(), article5Text);
+		System.out.println("Verifying Attendee Article 5 Url: " + joinHelppg.attendeeHelpLinks.get(4).getAttribute("href") );
 		Assert.assertTrue(joinHelppg.attendeeHelpLinks.get(4).getAttribute("href").
 					contains(article5Url));
 		joinHelppg.attendeeHelpLinks.get(4).click();
 		driver.navigate().back();
+		joinHelppg.imAttendee.click();
 		
 		//Verify the 6th Article
+		System.out.println("Verifying Attendee Article 6 Text:"+ joinHelppg.attendeeHelpLinks.get(5).getText());
 		Assert.assertEquals(joinHelppg.attendeeHelpLinks.get(5).getText(), article6Text);
+		System.out.println("Verifying Attendee Article 6 Url: " + joinHelppg.attendeeHelpLinks.get(5).getAttribute("href") );
 		Assert.assertTrue(joinHelppg.attendeeHelpLinks.get(5).getAttribute("href").
 					contains(article6Url));
 		joinHelppg.attendeeHelpLinks.get(5).click();
 		driver.navigate().back();
+		joinHelppg.imAttendee.click();
+		
 		}
 		catch(Exception ex){
 			System.out.println("Something went wrong in the verifyAttendeePage()");
@@ -136,12 +158,12 @@ public class Test_JoinHelp_Page {
 		*/
 	}
 	
-	@Test
+	@Test //(enabled = false)
 	public void verifyOrganizerPage(){
 		
 		String article1Text = "I’m having trouble downloading the software";
-		String article1Url = "http://support.citrixonline.com/Meeting/all_files/"
-				+ "G2M050002#Download?__";
+		String article1Url = "http://support.citrixonline.com/en_US/Meeting/all_files/G2M030004?__";
+		
 		
 		String article2Text = "I’m having trouble connecting online using the Web App.";
 		String article2Url = "http://support.citrixonline.com/en_US/Meeting/all_files/G2M090008?__";
@@ -158,52 +180,86 @@ public class Test_JoinHelp_Page {
 		String article6Text = "See other topics.";
 		String article6Url = "http://help.citrix.com/meeting";
 				
-				joinHelppg.imOrganizer.click();
+		joinHelppg.imOrganizer.click();
 	
 		//Verify the 1st Article
-		System.out.println("Verifying Article 1 Text:"+ joinHelppg.organizerHelpLinks.get(0).getText());
+		System.out.println("Verifying Organizer Article 1 Text:"+ joinHelppg.organizerHelpLinks.get(0).getText());
 		Assert.assertEquals(joinHelppg.organizerHelpLinks.get(0).getText(), article1Text);
-		System.out.println("Verifying Article 1 Url: " + joinHelppg.organizerHelpLinks.get(0).getAttribute("href") );
+		System.out.println("Verifying Organizer Article 1 Url: " + joinHelppg.organizerHelpLinks.get(0).getAttribute("href") );
 		Assert.assertTrue(joinHelppg.organizerHelpLinks.get(0).getAttribute("href").
 					contains(article1Url));
+			joinHelppg.organizerHelpLinks.get(0).click();
+			driver.navigate().back();
+			joinHelppg.imOrganizer.click();
 		
 		//Verify the 2nd Article
-		System.out.println("Verifying Article 2 Text:"+ joinHelppg.organizerHelpLinks.get(1).getText());
+		System.out.println("Verifying Organizer Article 2 Text:"+ joinHelppg.organizerHelpLinks.get(1).getText());
 		Assert.assertEquals(joinHelppg.organizerHelpLinks.get(1).getText(), article2Text);
 		
-		System.out.println("Verifying Article 2 Url: " + joinHelppg.organizerHelpLinks.get(1).getAttribute("href") );
+		System.out.println("Verifying Organizer Article 2 Url: " + joinHelppg.organizerHelpLinks.get(1).getAttribute("href") );
 		Assert.assertTrue(joinHelppg.organizerHelpLinks.get(1).getAttribute("href").
 					contains(article2Url));
+		joinHelppg.organizerHelpLinks.get(1).click();
+		driver.navigate().back();
+		joinHelppg.imOrganizer.click();
+	
 		
 		//Verify the 3rd Article
-		System.out.println("Verifying Article 3 Text:"+ joinHelppg.organizerHelpLinks.get(2).getText());
+		System.out.println("Verifying Organizer Article 3 Text:"+ joinHelppg.organizerHelpLinks.get(2).getText());
 		Assert.assertEquals(joinHelppg.organizerHelpLinks.get(2).getText(), article3Text);
-		System.out.println("Verifying Article 3 Url: " + joinHelppg.organizerHelpLinks.get(2).getAttribute("href") );
+		System.out.println("Verifying Organizer Article 3 Url: " + joinHelppg.organizerHelpLinks.get(2).getAttribute("href") );
 		Assert.assertTrue(joinHelppg.organizerHelpLinks.get(2).getAttribute("href").
 					contains(article3Url));
+		joinHelppg.organizerHelpLinks.get(2).click();
+		driver.navigate().back();
+		joinHelppg.imOrganizer.click();
+	
 		
 		//Verify the 4th Article
-		System.out.println("Verifying Article 4 Text:"+ joinHelppg.organizerHelpLinks.get(3).getText());
+		System.out.println("Verifying Organizer Article 4 Text:"+ joinHelppg.organizerHelpLinks.get(3).getText());
 		Assert.assertEquals(joinHelppg.organizerHelpLinks.get(3).getText(), article4Text);
-		System.out.println("Verifying Article 4 Url: " + joinHelppg.organizerHelpLinks.get(3).getAttribute("href") );
+		System.out.println("Verifying Organizer Article 4 Url: " + joinHelppg.organizerHelpLinks.get(3).getAttribute("href") );
 		Assert.assertTrue(joinHelppg.organizerHelpLinks.get(3).getAttribute("href")
 				.contains(article4Url));
-				
+		joinHelppg.organizerHelpLinks.get(3).click();
+		driver.navigate().back();
+		joinHelppg.imOrganizer.click();
+		
 		//Verify the 5th Article
-		System.out.println("Verifying Article 5 Text:"+ joinHelppg.organizerHelpLinks.get(4).getText());
+		System.out.println("Verifying Organizer Article 5 Text:"+ joinHelppg.organizerHelpLinks.get(4).getText());
 		Assert.assertEquals(joinHelppg.organizerHelpLinks.get(4).getText(), article5Text);
-		System.out.println("Verifying Article 5 Url: " + joinHelppg.organizerHelpLinks.get(4).getAttribute("href") );
+		System.out.println("Verifying Organizer Article 5 Url: " + joinHelppg.organizerHelpLinks.get(4).getAttribute("href") );
 		Assert.assertTrue(joinHelppg.organizerHelpLinks.get(4).getAttribute("href").
 					contains(article5Url));		
+		joinHelppg.organizerHelpLinks.get(4).click();
+		driver.navigate().back();
+		joinHelppg.imOrganizer.click();
 		
 		//Verify the 6th Article
-		System.out.println("Verifying Article 6 Text:"+ joinHelppg.organizerHelpLinks.get(5).getText());
+		System.out.println("Verifying Organizer Article 6 Text:"+ joinHelppg.organizerHelpLinks.get(5).getText());
 		Assert.assertEquals(joinHelppg.organizerHelpLinks.get(5).getText(), article6Text);
-		System.out.println("Verifying Article 6 Url: " + joinHelppg.organizerHelpLinks.get(5).getAttribute("href") );
+		System.out.println("Verifying Organizer Article 6 Url: " + joinHelppg.organizerHelpLinks.get(5).getAttribute("href") );
 		Assert.assertTrue(joinHelppg.organizerHelpLinks.get(5).getAttribute("href").
-					contains(article6Url));				
+					contains(article6Url));		
+		joinHelppg.organizerHelpLinks.get(5).click();
+		driver.navigate().back();
+		joinHelppg.imOrganizer.click();
 		
 		
+	}
+	
+	@AfterTest
+	public void closeBrowser(){
+		try{
+			driver.quit();
+		}
+		catch(Exception e){
+			System.out.println("Exception in trying to close Browser: " + e.toString());
+			throw(e);
+		}
+		finally{
+			
+		}
 	}
